requirements.txt
arabic-reshaper
argon2-cffi
argon2-cffi-bindings
asgiref
asn1crypto
Brotli
certifi
cffi
chardet
charset-normalizer
click
cryptography
cssselect2
defusedxml
Django
django-autocomplete-light
django-extensions
django-rangefilter
django-smtp-ssl
django-widget-tweaks
djangorestframework
et_xmlfile
fonttools
gunicorn
html5lib
idna
lxml
mysqlclient
numpy
openpyxl
oscrypto
pandas
pdfkit
pillow
pycparser
pydyf
pyHanko
pyhanko-certvalidator
pypdf
pyphen
python-bidi
python-dateutil
python-decouple
pytz
PyYAML
qrcode
reportlab
requests
six
sqlparse
svglib
tinycss2
tinyhtml5
typing_extensions
tzdata
tzlocal
uritools
urllib3
weasyprint
webencodings
xhtml2pdf
zopfli
pymysql
whitenoise
django-ckeditor
drf-yasg